LPS-57354 Do not use hierarchy as method library
