+++ modules/apps/blogs/blogs-web/src/com/liferay/blogs/web/image/selector/BaseBlogsImageSelectorUploadHandler.java
@@ -20,12 +20,10 @@ import com.liferay.portal.kernel.image.selector.BaseImageSelectorUploadHandler;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.portlet.JSONPortletResponseUtil;
-import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.servlet.ServletResponseConstants;
 import com.liferay.portal.kernel.util.FileUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.TempFileEntryUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.security.permission.ActionKeys;
@@ -40,7 +38,6 @@ import com.liferay.portlet.documentlibrary.FileNameException;
 import com.liferay.portlet.documentlibrary.antivirus.AntivirusScannerException;
 
 import java.io.IOException;
-import java.io.InputStream;
 
 import javax.portlet.PortletRequest;
 import javax.portlet.PortletResponse;
@@ -73,6 +70,10 @@ public abstract class BaseBlogsImageSelectorUploadHandler
 	public void validateFile(String fileName, String contentType, long size)
 		throws PortalException {
 
+		if (size > getMaxFileSize()) {
+			throw new BlogImageSizeException();
+		}
+
 		String extension = FileUtil.getExtension(fileName);
 
 		String[] imageExtensions = PrefsPropsUtil.getStringArray(
@@ -87,36 +88,12 @@ public abstract class BaseBlogsImageSelectorUploadHandler
 		}
 
 		throw new BlogImageNameException(
-			"Invalid cover image for file name " + fileName);
+			"Invalid image for file name " + fileName);
 	}
 
-	@Override
-	protected FileEntry addFileEntry(
-			ThemeDisplay themeDisplay, String fileName, InputStream inputStream,
-			String contentType)
-		throws PortalException {
-
-		return TempFileEntryUtil.addTempFileEntry(
-			themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
-			_TEMP_FOLDER_NAME, fileName, inputStream, contentType);
+	protected long getMaxFileSize() {
+		return PrefsPropsUtil.getLong(PropsKeys.BLOGS_IMAGE_MAX_SIZE);
 	}
-
-	@Override
-	protected FileEntry fetchFileEntry(
-			ThemeDisplay themeDisplay, String fileName)
-		throws PortalException {
-
-		try {
-			return TempFileEntryUtil.getTempFileEntry(
-				themeDisplay.getScopeGroupId(), themeDisplay.getUserId(),
-				_TEMP_FOLDER_NAME, fileName);
-		}
-		catch (PortalException pe) {
-			return null;
-		}
-	}
-
-	protected abstract long getMaxFileSize();
 
 	@Override
 	protected void handleUploadException(
